Add arity-specific callback handlers and generic PushHandler
- Add CallbackHandler1<A>, CallbackHandler2<A,B>, CallbackHandler3<A,B,C>
  functional interfaces for typed callback parameters
- Add corresponding registerCallback overloads to PythonEmbed and
  PythonEmbedPool (arity 1, 2, 3) delegating to existing typed registration
- Make PushHandler generic (PushHandler<T>) with throws Exception,
  replacing PushHandler1<T> and eliminating lambda ambiguity
- Update registerPushHandler overloads in PythonEmbed and PythonEmbedPool
  to use generic PushHandler<T> with exception-safe wrapping
- Remove (PushHandler) casts from tests, enabling natural syntax:
  py.registerPushHandler(\\log\\, Integer.class, (name, value) -> ...)

+++ python-embed-runtime/src/test/java/io/github/howtis/pythonembed/PythonEmbedCallbackTest.java
@@ -146,7 +146,7 @@ class PythonEmbedCallbackTest {
     void push_typedValue_noDowncasting() throws Exception {
         AtomicInteger received = new AtomicInteger(0);
         py.registerPushHandler("typed_progress", Integer.class,
-                (name, value) -> received.set((Integer) value));
+                (name, value) -> received.set(value));
 
         py.exec("_bridge.push('typed_progress', 75)");
         Thread.sleep(200);
@@ -158,7 +158,7 @@ class PythonEmbedCallbackTest {
     void push_doesNotBlockPython() {
         AtomicInteger received = new AtomicInteger(0);
         py.registerPushHandler("log", Integer.class, (name, value) -> {
-            received.set((Integer) value);
+            received.set(value);
         });
 
         // Push followed by computation should work
